pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepmind-workspace"
version = "1.0.0"
description = "Modular AI Assistant with Persistent Context, Document Integration, and DeepSeek API"
requires-python = ">=3.11"
dependencies = [
    "nicegui>=2.10.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "chromadb>=0.5.23",
    "sentence-transformers>=3.4.0",
    "tiktoken>=0.8.0",
    "pygithub>=2.5.0",
    "dropbox>=12.0.2",
    "google-api-python-client>=2.160.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "watchfiles>=1.0.3",
    "python-dotenv>=1.0.1",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "markdown>=3.7",
    "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
deepmind = "deepmind.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/deepmind --cov-report=term-missing"
